requirements.txt
Django==4.2
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
psycopg2==2.9.10
pulsar-client==3.5.0
celery==5.4.0
redis==5.2.1